% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_spline.R
\name{fit_spline}
\alias{fit_spline}
\title{Fit spline model to simulation data}
\usage{
fit_spline(sim_data, target)
}
\arguments{
\item{target}{}
}
\value{

}
\description{
Fit spline model to simulation data
}
