% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlcali-package.R
\docType{package}
\name{nlcali-package}
\alias{nlcali-package}
\alias{nlcali}
\title{The 'nlcali' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2022). RStan: the R interface to Stan. R package version 2.21.5. https://mc-stan.org
}
