% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{plot_res_2d}
\alias{plot_res_2d}
\title{Plot model fit with 1 baseline observation}
\usage{
plot_res_2d(output)
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Plot model fit with 1 baseline observation
}
