% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Run simulations for a range of starting_EIR values}
\usage{
run_simulations(parameters, target, target_tt, test_EIRs, ncores)
}
\value{

}
\description{
Run simulations for a range of starting_EIR values
}
